+++ java/CHANGELOG.md
@@ -1,6 +1,7 @@
 ### 0.9.3
 * Add Apache Sling DavEx Servlet OSGI config support #11
 * Fix missing getLoginPage operation response body
+* Add sling get agents support #16
 
 ### 0.9.2
 * Add operation parameter support to sling#postAgentWithHttpInfo